Added initial class that represents state
The class represents the current state of the town/village. The move method takes care of updating parcels with the parcel the robot is carrying (parcels with the same `current` position as the robot itself), and if the destination is equal to the address of the parcel, it is filtered (hence delivering the parcel), and the state is updated. This code currently contains no automation.